refactor: extract findFunctionNode to shared utility
- Remove duplicated findFunctionNode methods from NamingQualityAnalyzer and TypeSafetyAnalyzer
- Create shared ast-utils.ts with proper documentation
- Reduces code duplication and complexity (CC: 13 → 0 for both instances)
